+++ src/containers.c
@@ -139,7 +139,7 @@ StrToIntArray(const char *string, IntArray **int_array_ptr)
    free(buffer);
 
    /* Create IntArray */
-   int_array = IntArrayCreate(count);
+   int_array = IntArrayCreate((size_t)count);
 
    /* Build array */
    buffer = strdup(string);
@@ -181,7 +181,7 @@ StrToStackIntArray(const char *string, StackIntArray *int_array)
 
    /* Set StackIntArray size */
    int_array->size =
-      (count < MAX_STACK_ARRAY_LENGTH) ? count : MAX_STACK_ARRAY_LENGTH - 1;
+      (count < MAX_STACK_ARRAY_LENGTH) ? (size_t)count : MAX_STACK_ARRAY_LENGTH - 1;
 
    /* Build array */
    buffer = strdup(string);
@@ -266,7 +266,7 @@ IntArrayUnique(MPI_Comm comm, IntArray *int_array)
    /* Gather sizes of local unique arrays */
    if (!myid)
    {
-      all_num_entries = (int *)malloc(nprocs * sizeof(int));
+      all_num_entries = (int *)malloc((size_t)nprocs * sizeof(int));
    }
    num_entries_int = (int)int_array->unique_size;
    MPI_Gather(&num_entries_int, 1, MPI_INT, all_num_entries, 1, MPI_INT, 0, comm);
@@ -274,14 +274,14 @@ IntArrayUnique(MPI_Comm comm, IntArray *int_array)
    /* Gather local unique arrays */
    if (!myid)
    {
-      displs            = (int *)calloc(nprocs, sizeof(int));
+      displs            = (int *)calloc((size_t)nprocs, sizeof(int));
       total_num_entries = all_num_entries[0];
       for (int i = 1; i < nprocs; i++)
       {
          displs[i] = displs[i - 1] + all_num_entries[i - 1];
          total_num_entries += all_num_entries[i];
       }
-      all_data = (int *)malloc(total_num_entries * sizeof(int));
+      all_data = (int *)malloc((size_t)total_num_entries * sizeof(int));
    }
    MPI_Gatherv(int_array->unique_data, (int)int_array->unique_size, MPI_INT, all_data,
                all_num_entries, displs, MPI_INT, 0, comm);
@@ -290,7 +290,7 @@ IntArrayUnique(MPI_Comm comm, IntArray *int_array)
    if (!myid)
    {
       /* Sort input array */
-      qsort(all_data, total_num_entries, sizeof(int), IntArrayCompare);
+      qsort(all_data, (size_t)total_num_entries, sizeof(int), IntArrayCompare);
 
       /* Find number of unique entries */
       int_array->g_unique_size = 1;
@@ -355,7 +355,7 @@ IntArrayParRead(MPI_Comm comm, const char *prefix, IntArray **int_array_ptr)
    }
 
    /* 1b) Compute partids array */
-   partids = malloc(nparts * sizeof(int));
+   partids = malloc((size_t)nparts * sizeof(int));
    offset  = myid * nparts;
    offset += (myid < (g_nparts % nprocs)) ? myid : (g_nparts % nprocs);
    for (int part = 0; part < nparts; part++)
@@ -465,8 +465,8 @@ IntArrayBuild(MPI_Comm comm, int size, const int *dofmap, IntArray **int_array_p
 {
    IntArray *int_array = NULL;
 
-   int_array = IntArrayCreate(size);
-   memcpy(int_array->data, dofmap, size * sizeof(int));
+   int_array = IntArrayCreate((size_t)size);
+   memcpy(int_array->data, dofmap, (size_t)size * sizeof(int));
    IntArrayUnique(comm, int_array);
 
    *int_array_ptr = int_array;
@@ -483,7 +483,7 @@ IntArrayBuildInterleaved(MPI_Comm comm, int num_local_blocks, int num_dof_types,
    IntArray *int_array = NULL;
    int       size      = num_dof_types * num_local_blocks; // TODO: check overflow
 
-   int_array = IntArrayCreate(size);
+   int_array = IntArrayCreate((size_t)size);
    for (int i = 0; i < num_local_blocks; i++)
    {
       for (int j = 0; j < num_dof_types; j++)
@@ -507,7 +507,7 @@ IntArrayBuildContiguous(MPI_Comm comm, int num_local_blocks, int num_dof_types,
    IntArray *int_array = NULL;
    int       size      = num_dof_types * num_local_blocks; // TODO: check overflow
 
-   int_array = IntArrayCreate(size);
+   int_array = IntArrayCreate((size_t)size);
    for (int i = 0; i < num_dof_types; i++)
    {
       for (int j = 0; j < num_local_blocks; j++)
